# golang/openAPI.json
{"openapi":"3.0.0","servers":[{"url":"http://example.com","description":"","variables":{}}],"info":{"version":"2de4cabf-0659-4851-98f5-89007145b425","title":"efishery","description":"","termsOfService":"","contact":{},"license":{"name":""}},"paths":{"/register":{"post":{"summary":"1. Create User","operationId":"1.CreateUser","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"roleId":2,"name":"riski","phone":"082128826451"}},"example":"{\n    \"roleId\" : 2,\n    \"name\" : \"riski\",\n    \"phone\" :\"082128826451\"\n}"}}},"tags":["Auth"]}},"/login":{"post":{"summary":"2. Login","operationId":"2.Login","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":"{\n    \"phone\" : \"082128826458\",\n    \"password\" : \"JhBa\"\n}\n/*Admin*/\n/*\n     \"phone\" : \"081212341234\",\n     \"password\" : \"jnmp\"\n} */\n/*Guest*/"},"example":"{\n    \"phone\" : \"082128826458\",\n    \"password\" : \"JhBa\"\n}\n/*Admin*/\n/*\n     \"phone\" : \"081212341234\",\n     \"password\" : \"jnmp\"\n} */\n/*Guest*/"}}},"tags":["Auth"]}},"/verify":{"post":{"summary":"3. Verify JWT","operationId":"3.VerifyJWT","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MTI2MDcwNDcsImlhdCI6MTYxMjU2Mzg0NywibmFtZSI6Imd1ZXN0IiwicGhvbmUiOiIwODIxMjg4MjY0NTgiLCJyb2xlSWQiOjIsInRpbWVzdGFtcCI6IjIwMjEtMDItMDZUMTc6MjQ6MDcuNTc2MDU2KzA3OjAwIn0.5B1OBN1p2NwstrNNIOUbW0Gf-qMcSRmboXvwG6fjNec"}},"example":"{\n    \"token\" : \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MTI2MDcwNDcsImlhdCI6MTYxMjU2Mzg0NywibmFtZSI6Imd1ZXN0IiwicGhvbmUiOiIwODIxMjg4MjY0NTgiLCJyb2xlSWQiOjIsInRpbWVzdGFtcCI6IjIwMjEtMDItMDZUMTc6MjQ6MDcuNTc2MDU2KzA3OjAwIn0.5B1OBN1p2NwstrNNIOUbW0Gf-qMcSRmboXvwG6fjNec\"\n}"}}},"tags":["Auth"]}},"/auth/users":{},"/auth/prices":{"get":{"summary":"1. List data Price","operationId":"1.ListdataPrice","parameters":[],"responses":{"200":{"description":"","headers":{}}},"tags":["Fetch"]}},"/auth/admin/aggregate":{"get":{"summary":"2. Aggregate Data","operationId":"2.AggregateData","parameters":[{"$ref":"#/components/parameters/dateFrom"},{"$ref":"#/components/parameters/dateTo"},{"$ref":"#/components/parameters/areaProvinsi"}],"responses":{"200":{"description":"","headers":{}}},"tags":["Fetch"]}},"/auth/verify":{"post":{"summary":"3. Verify JWT","operationId":"3.VerifyJWT2","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MTI2MDcwNDcsImlhdCI6MTYxMjU2Mzg0NywibmFtZSI6Imd1ZXN0IiwicGhvbmUiOiIwODIxMjg4MjY0NTgiLCJyb2xlSWQiOjIsInRpbWVzdGFtcCI6IjIwMjEtMDItMDZUMTc6MjQ6MDcuNTc2MDU2KzA3OjAwIn0.5B1OBN1p2NwstrNNIOUbW0Gf-qMcSRmboXvwG6fjNec"}},"example":"{\n    \"token\" : \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MTI2MDcwNDcsImlhdCI6MTYxMjU2Mzg0NywibmFtZSI6Imd1ZXN0IiwicGhvbmUiOiIwODIxMjg4MjY0NTgiLCJyb2xlSWQiOjIsInRpbWVzdGFtcCI6IjIwMjEtMDItMDZUMTc6MjQ6MDcuNTc2MDU2KzA3OjAwIn0.5B1OBN1p2NwstrNNIOUbW0Gf-qMcSRmboXvwG6fjNec\"\n}"}}},"tags":["Fetch"]}}},"components":{"securitySchemes":{"noauth":{"type":"http","scheme":"noauth"},"bearer":{"type":"http","scheme":"bearer"}},"parameters":{"page":{"name":"page","in":"query","required":true,"style":"form","schema":{"type":"number"}},"limiit":{"name":"limiit","in":"query","required":true,"style":"form","schema":{"type":"number"}},"name":{"name":"name","in":"query","required":true,"style":"form","schema":{"type":"string","example":"ad"}},"email":{"name":"email","in":"query","required":true,"style":"form","schema":{"type":"number"}},"dateFrom":{"name":"dateFrom","in":"query","required":true,"style":"form","schema":{"type":"string","example":"2020-07-23"}},"dateTo":{"name":"dateTo","in":"query","required":true,"style":"form","schema":{"type":"string","example":"2020-07-30"}},"areaProvinsi":{"name":"areaProvinsi","in":"query","required":true,"style":"form","schema":{"type":"string","example":"banten"}}}},"security":[{"noauth":[]},{"bearer":[]},{"bearer":[]},{"bearer":[]},{"bearer":[]}],"tags":[],"externalDocs":{"url":"","description":""},"warnings":[]}